refactor(admin): Q9.d — extract admin/mcp_templates.rs (GAR-472)
Move McpTemplate struct + builtin_templates (private) + load/save helpers
(private) + list_mcp_templates + save_mcp_template + delete_mcp_template
from admin/handlers.rs (2550 LOC) to new admin/mcp_templates.rs (244 LOC).

handlers.rs: 2550 → 2326 LOC (−224). Zero behavior change.
Re-exported via `pub use super::mcp_templates::*` so routes.rs paths
(`handlers::list_mcp_templates`, etc.) keep resolving without changes.

Follows exact pattern of Q9.b (GAR-470/providers.rs) and Q9.c (GAR-471/mcp.rs).
