+++ crates/garraia-gateway/src/admin/mod.rs
@@ -1,6 +1,7 @@
 pub mod audit;
 pub mod handlers;
 pub mod mcp;
+pub mod mcp_templates;
 pub mod middleware;
 pub mod providers;
 pub mod rbac;
